requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
av==15.0.0
certifi==2025.4.26
charset-normalizer==2.0.12
datasets==3.5.0
dill==0.3.8
filelock==3.13.1
frozenlist==1.6.0
fsspec==2024.6.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
pyarrow==20.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==70.2.0
shortuuid==1.0.13
six==1.17.0
sympy==1.13.3
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
xxhash==3.5.0
yarl==1.20.0
